@@ -20,6 +20,6 @@
     "jest": "^27.3.1",
     "ts-jest": "^27.0.7",
     "tsd": "^0.18.0",
-    "typescript": "^4.4.4"
+    "typescript": "^4.9.5"
   }
 }
