@@ -17,40 +17,35 @@
   },
   "license": "MIT",
   "scripts": {
-    "generate:example": "ts-node -r tsconfig-paths/register example.ts",
-    "cli:example": "cd example && node ../dist/src/cli.js",
-    "precli:example": "yarn build",
+    "generate:example": "ts-node --project ./tsconfig.src.json -r tsconfig-paths/register example.ts",
+    "cli:example": "yarn build && cd example && node ../dist/src/cli.js",
     "typecheck": "tsc -p . --noEmit",
     "lint": "run-s lint:*",
     "lint:eslint": "eslint . --ext ts",
-    "lint:prettier": "yarn prettier --check",
+    "lint:prettier": "yarn run:prettier --check",
     "fix": "run-s fix:*",
     "fix:eslint": "yarn lint:eslint --fix",
-    "fix:prettier": "yarn prettier --write",
-    "prettier": "prettier '**/*.{ts,js,json,md}'",
-    "build": "tsc -p . && tsc-alias -p tsconfig.json",
+    "fix:prettier": "yarn run:prettier --write",
+    "run:prettier": "prettier '**/*.{ts,mts,mcs,mjs,cjs,js,json,md,yml,yaml}'",
+    "build": "tsc -p . && tsc-alias -p tsconfig.src.json",
     "test": "jest",
     "test:watch": "jest --watch",
     "lint-staged": "lint-staged --allow-empty",
     "prepare": "husky install",
-    "prepublishOnly": "yarn build",
-    "patch": "patch-package"
+    "prepublishOnly": "yarn build"
   },
-  "lint-staged": {
-    "**/*.ts": [
-      "eslint --fix"
-    ],
-    "**/*.{ts,js,json,md}": [
-      "prettier --write"
-    ]
+  "peerDependencies": {
+    "typescript": "4.4.4"
   },
   "dependencies": {
     "commander": "^8.3.0",
     "glob": "^7.2.0",
-    "ramda": "^0.27.1",
-    "typescript": "^4.4.4"
+    "ramda": "^0.27.1"
   },
   "devDependencies": {
+    "@swc/core": "^1.3.32",
+    "@swc/jest": "^0.2.24",
+    "@tsconfig/node18-strictest": "^1.0.0",
     "@types/dedent": "^0.7.0",
     "@types/glob": "^7.2.0",
     "@types/jest": "^27.0.2",
@@ -66,14 +61,14 @@
     "jest": "^27.3.1",
     "lint-staged": "^12.0.2",
     "npm-run-all": "^4.1.5",
-    "patch-package": "^6.4.7",
     "postinstall-postinstall": "^2.1.0",
     "prettier": "^2.4.1",
     "ts-jest": "^27.0.7",
     "ts-node": "^10.4.0",
     "tsc-alias": "^1.4.0",
     "tsconfig-paths": "^3.11.0",
-    "tsd": "^0.18.0"
+    "tsd": "^0.18.0",
+    "typescript": "^4.9.5"
   },
   "keywords": [
     "typescript",
