@@ -1,11 +1,7 @@
-const { pathsToModuleNameMapper } = require("ts-jest/utils")
-const {
-  readConfigFile,
-  parseJsonConfigFileContent,
-  sys,
-} = require("typescript")
+import { pathsToModuleNameMapper } from "ts-jest"
+import { readConfigFile, parseJsonConfigFileContent, sys } from "typescript"
 
-const configFile = readConfigFile("./tsconfig.json", sys.readFile)
+const configFile = readConfigFile("./tsconfig.src.json", sys.readFile)
 if (typeof configFile.error !== "undefined") {
   throw new Error(`Failed to load tsconfig: ${configFile.error}`)
 }
@@ -23,17 +19,17 @@ const { options } = parseJsonConfigFileContent(
 
 module.exports = {
   roots: ["<rootDir>"],
-  testMatch: ["**/tests/**/?(*.)+(spec|test).+(ts|tsx|js)"],
+  testMatch: ["**/?(*.)+(spec|test).+(ts|tsx)"],
   testPathIgnorePatterns: ["<rootDir>/example/"],
   globals: {
     "ts-jest": {
-      tsconfig: "tests/tsconfig.json",
+      tsconfig: "tsconfig.test.json",
     },
   },
   transform: {
-    "^.+\\.(t|j)s$": "ts-jest",
+    "^.+\\.(t|j)s$": "@swc/jest",
   },
-  moduleNameMapper: pathsToModuleNameMapper(options.paths, {
+  moduleNameMapper: pathsToModuleNameMapper(options.paths ?? {}, {
     prefix: "<rootDir>",
   }),
 }
