@@ -1,7 +1,7 @@
 module.exports = {
   parser: "@typescript-eslint/parser",
   parserOptions: {
-    project: "./**/tsconfig.json",
+    project: "./**/(tsconfig.src|tsconfig.test).json",
     sourceType: "module",
   },
   extends: [
